Clean Architecture 적용
