+++ app/src/main/java/han/ayeon/searchimgandvideo/domain/data/FetchApiResult.kt
@@ -1,4 +1,4 @@
-package han.ayeon.searchimgandvideo.model.data
+package han.ayeon.searchimgandvideo.domain.data
 
 /**
  * Created by HanAYeon on 2019-07-17.
