+++ app/src/main/java/han/ayeon/searchimgandvideo/domain/repository/searchlist/ImageRepository.kt
@@ -1,0 +1,12 @@
+package han.ayeon.searchimgandvideo.domain.repository.searchlist
+
+import io.reactivex.Observable
+
+/**
+ * Created by HanAYeon on 2019-07-23.
+ */
+
+interface ImageRepository {
+
+    fun get(searchWord : String) : Observable<Any>?
+}
