+++ app/src/main/java/han/ayeon/searchimgandvideo/domain/repository/searchlist/VideoRepository.kt
@@ -1,0 +1,12 @@
+package han.ayeon.searchimgandvideo.domain.repository.searchlist
+
+import han.ayeon.searchimgandvideo.domain.data.VideoResponse
+import io.reactivex.Observable
+
+/**
+ * Created by HanAYeon on 2019-07-23.
+ */
+
+interface VideoRepository {
+    fun get(searchWord : String) : Observable<Any>?
+}
