--- app/src/main/java/han/ayeon/searchimgandvideo/model/data/Meta.kt
@@ -1,11 +1,0 @@
-package han.ayeon.searchimgandvideo.model.data
-
-/**
- * Created by HanAYeon on 2019-07-17.
- */
-
-data class Meta(
-        val isEnd: Boolean = false,
-        val pageableCount: Int = 0,
-        val totalCount: Int = 0
-)
