--- app/src/main/java/han/ayeon/searchimgandvideo/model/data/VideoDocument.kt
@@ -1,27 +1,0 @@
-package han.ayeon.searchimgandvideo.model.data
-
-import com.google.gson.annotations.SerializedName
-
-/**
- * Created by HanAYeon on 2019-07-17.
- */
-
-data class VideoDocument(
-        @SerializedName("author")
-        val author: String,
-
-        @SerializedName("datetime")
-        val datetime: String,
-
-        @SerializedName("play_time")
-        val playTime: Int,
-
-        @SerializedName("thumbnail")
-        val thumbnail: String,
-
-        @SerializedName("title")
-        val title: String,
-
-        @SerializedName("url")
-        val url: String
-)
