--- app/src/main/java/han/ayeon/searchimgandvideo/model/data/VideoResponse.kt
@@ -1,10 +1,0 @@
-package han.ayeon.searchimgandvideo.model.data
-
-/**
- * Created by HanAYeon on 2019-07-17.
- */
-
-data class VideoResponse(
-        val meta: Meta,
-        val documents: List<VideoDocument>
-)
