--- app/src/main/java/han/ayeon/searchimgandvideo/model/web/SearchService.kt
@@ -1,12 +1,0 @@
-package han.ayeon.searchimgandvideo.model.web
-
-import han.ayeon.searchimgandvideo.model.data.FetchMediaApiResult
-
-/**
- * Created by HanAYeon on 2019-07-17.
- */
-
-interface SearchService {
-
-    fun search(searchWord : String, queryCallBack : FetchMediaApiResult)
-}
