+++ app/src/main/java/han/ayeon/searchimgandvideo/presentation/view/ImageLoader.kt
@@ -1,4 +1,4 @@
-package han.ayeon.searchimgandvideo.view
+package han.ayeon.searchimgandvideo.presentation.view
 
 /**
  * Created by HanAYeon on 2019-07-18.
