+++ app/src/main/java/han/ayeon/searchimgandvideo/presentation/view/savedlist/SavedListFragment.kt
@@ -1,4 +1,4 @@
-package han.ayeon.searchimgandvideo.view.savedlist
+package han.ayeon.searchimgandvideo.presentation.view.savedlist
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -11,8 +11,8 @@ import androidx.lifecycle.ViewModelProviders
 import androidx.recyclerview.widget.GridLayoutManager
 import han.ayeon.searchimgandvideo.R
 import han.ayeon.searchimgandvideo.databinding.FragmentResultListBinding
-import han.ayeon.searchimgandvideo.model.data.Media
-import han.ayeon.searchimgandvideo.viewmodel.SearchResultViewModel
+import han.ayeon.searchimgandvideo.domain.data.Media
+import han.ayeon.searchimgandvideo.presentation.viewmodel.SearchResultViewModel
 
 /**
  * Created by HanAYeon on 2019-07-18.
@@ -42,10 +42,10 @@ class SavedListFragment : Fragment() {
         fragmentResultListBinding.imageListView.adapter = viewAdapter
 
 
-        viewModel.savedList.observe( this,  object : Observer<ArrayList<Media>>{
+        viewModel.getSavedListLiveData().observe( this,  object : Observer<ArrayList<Media>>{
             override fun onChanged(t: ArrayList<Media>?) {
-                if(viewModel.savedList.value == null) return
-                savedListChange(viewModel.savedList.value!!)
+                if(viewModel.getSavedItem() == null) return
+                savedListChange(viewModel.getSavedItem()!!)
             }
 
         })
