+++ app/src/main/java/han/ayeon/searchimgandvideo/presentation/view/searchlist/SearchListViewHolder.kt
@@ -1,12 +1,12 @@
-package han.ayeon.searchimgandvideo.view.searchlist
+package han.ayeon.searchimgandvideo.presentation.view.searchlist
 
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import han.ayeon.searchimgandvideo.R
 import han.ayeon.searchimgandvideo.databinding.FragmentSearchResultItemBinding
-import han.ayeon.searchimgandvideo.model.data.Media
-import han.ayeon.searchimgandvideo.view.ImageLoader
-import han.ayeon.searchimgandvideo.viewmodel.SearchResultViewModel
+import han.ayeon.searchimgandvideo.domain.data.Media
+import han.ayeon.searchimgandvideo.presentation.view.ImageLoader
+import han.ayeon.searchimgandvideo.presentation.viewmodel.SearchResultViewModel
 import kotlinx.android.synthetic.main.fragment_search_result_item.view.*
 
 /**
